+++ index.js
@@ -1,2 +1,3 @@
 import './modules/books.js';
 import './modules/time.js';
+import './modules/navigation.js';
