+++ package.json
@@ -16,5 +16,8 @@
   "bugs": {
     "url": "https://github.com/ismailmunyentwari9/book-ES6/issues"
   },
-  "homepage": "https://github.com/ismailmunyentwari9/book-ES6#readme"
+  "homepage": "https://github.com/ismailmunyentwari9/book-ES6#readme",
+  "devDependencies": {
+    "hint": "^7.1.3"
+  }
 }
